v2

+++ RebusNeo/Src/Domain/Implementations/Order.cs
@@ -7,7 +7,7 @@ namespace RebusNeo.Src.Domain.Implementations
     {
         public int id { get; set; }
         public int userid { get; set; }
-        public int cost { get; set; }
+        public decimal cost { get; set; }
         public string details { get; set; }
         public DateTime datetime { get; set; }
     }
