+++ RebusNeo/Src/Domain/Interfaces/IOrder.cs
@@ -6,7 +6,7 @@ namespace RebusNeo.Src.Domain.Interfaces
     {
         public int id { get; set; }
         public int userid { get; set; }
-        public int cost { get; set; }
+        public decimal cost { get; set; }
         public string details { get; set; }
         public DateTime datetime { get; set; }
     }
